reconfigurando projeto do zero
