--- semana17/aula49-filtros-ordenacao-paginacao/src/data/connection.ts
@@ -1,16 +1,0 @@
-import knex from "knex"
-import dotenv from "dotenv"
-
-dotenv.config()
-
-export const connection = knex({
-   client: "mysql",
-   connection: {
-      host: process.env.DB_HOST,
-      port: 3306,
-      user: process.env.DB_USER,
-      password: process.env.DB_PASSWORD,
-      database: process.env.DB_SCHEMA,
-      multipleStatements: true
-   }
-})
