--- semana17/aula49-filtros-ordenacao-paginacao/src/types.ts
@@ -1,15 +1,0 @@
-export type recipe = {
-   id: string
-   title: string
-   description: string
-   userId: string
-   createdAt: number
-}
-
-export type user = {
-   id: string
-   name: string
-   email: string
-   password: string
-   recipes?: recipe[]
-}
